[10.7][B][F] Secure APIs by adding API JWT verification
* All API now verify the the Role supplied in the JWT Token +
  to further protect against unauthorized access the user id stored in
  the JWT token is used.
* Removed `getUserFromID` API, since it leaks lots of information to
  everyone. Replaced it with `getUsernameFromUserID`.

+++ backend/src/main/java/controller/ListingController.java
@@ -3,8 +3,10 @@ package controller;
 import entity.Listing;
 import enums.Qualification;
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
+import io.quarkus.logging.Log;
 import io.quarkus.panache.common.Page;
 import io.quarkus.runtime.StartupEvent;
+import io.quarkus.security.UnauthorizedException;
 import jakarta.annotation.security.PermitAll;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.enterprise.event.Observes;
@@ -24,6 +26,7 @@ import org.hibernate.search.mapper.orm.scope.SearchScope;
 import org.hibernate.search.mapper.orm.session.SearchSession;
 import org.jboss.logging.Logger;
 import service.ListingService;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -36,6 +39,9 @@ import java.util.Set;
 public class ListingController {
     @Inject
     ListingService listingService;
+
+    @Inject
+    JsonWebToken jwt;
 
     @Inject
     SearchSession searchSession;
@@ -53,7 +59,7 @@ public class ListingController {
     }
 
     @Query("getAllListings")
-    @RolesAllowed({"ListingsConsumer", "ListingProvider"})
+    @RolesAllowed({"ListingConsumer", "ListingProvider", "Administrator"})
     @Description("Fetches a list of all listings from the database")
     public List<Listing> getAllListings() {
         LOG.info("getAllListings");
@@ -61,13 +67,20 @@ public class ListingController {
     }
 
     @Query("getAllListingsFromUserWithId")
+    @RolesAllowed({"ListingConsumer", "ListingProvider", "Administrator"})
     @Description("Fetches a list of all listings from the a specific User from the dataabase")
-    public List<Listing> getAllListingsFromUserWithId(long id) {
+    public List<Listing> getAllListingsFromUserWithId(Long id) throws GraphQLException{
         LOG.info("getAllListingsFromUserWithId");
+        // check if sender is allowed
+        if (!id.equals(Long.parseLong(jwt.getClaim("userid").toString())) && !jwt.getGroups().contains("Administrator")){
+            throw new GraphQLException("Not allowed");
+        }
+
         return listingService.getAllListingsFromUserWithId(id);
     }
 
     @Mutation
+    @RolesAllowed({"ListingProvider", "Administrator"})
     @Description("Creates a listing in the database")
     public Listing createListing(Listing listing) throws GraphQLException {
         LOG.info("createListing");
@@ -79,7 +92,7 @@ public class ListingController {
         }
     }
     @Mutation("applyToListing")
-    @PermitAll
+    @RolesAllowed({"ListingConsumer", "Administrator"})
     public void applyToListing(Long listingId, Long userId, String applicationText) throws GraphQLException {
         LOG.info("applyToListing");
         try {
@@ -91,7 +104,7 @@ public class ListingController {
     }
 
     @Query("advancedSearch")
-    @PermitAll
+    @RolesAllowed({"ListingConsumer","ListingProvider", "Administrator"})
     @Transactional
     public GraphQLSearchResult advancedSearch(Optional<String> textPattern, Optional<String> startDate, Optional<String> endDate,
                                               Optional<Qualification> qualification, Optional<String> university,
@@ -181,6 +194,7 @@ public class ListingController {
     }
 
     @Query("getTrendingListings")
+    @RolesAllowed({"ListingProvider", "ListingConsumer", "Administrator"})
     @Description("Fetches a list of trending listings from the database")
     public GraphQLSearchResult getTrendingListings(Optional<String> university, Optional<String> company,
                                                    Optional<Integer> pageIndex, Optional<Integer> pageSize) {
@@ -199,20 +213,26 @@ public class ListingController {
     }
 
 
+
     @Mutation
+    @RolesAllowed({"ListingProvider", "Administrator"})
     @Description("Updates a listing in the database")
     public Listing updateListing(Listing listing) throws GraphQLException {
-        LOG.info("updateListing");
-        LOG.info(listing);
+        // check if user (identified with JWT) is allowed to change listing
+        if (!is_user_allowed_to_access_listing(jwt,listing.getId()) && !jwt.getGroups().contains("Administrator")){
+            throw new UnauthorizedException();
+        }
         try {
+            //return listingService.getListingById(listing.getId());
             return listingService.updateListing(listing);
-        } catch (ValidationException | ServiceException e) {
+        } catch (ServiceException | ValidationException e) {
             LOG.error("Error in updateListing: " + e.getMessage());
             throw new GraphQLException(e.getMessage());
         }
     }
 
     @Query("getListingById")
+    @RolesAllowed({"ListingProvider", "ListingConsumer", "Administrator"})
     @Description("Fetches a listing from the database by its ID")
     public Listing getListingById(long id) throws GraphQLException {
         LOG.info("getListingById");
@@ -225,8 +245,13 @@ public class ListingController {
     }
 
     @Mutation
+    @RolesAllowed({"ListingProvider", "Administrator"})
     @Description("Deletes a listing and associated entities in cascade")
     public void deleteListingById(long id) throws GraphQLException {
+        // check if user (identified with JWT) is allowed to change listing
+        if (!is_user_allowed_to_access_listing(jwt, id) && !jwt.getGroups().contains("Administrator")){
+            throw new UnauthorizedException();
+        }
         LOG.info("deleteListingById");
         try {
             listingService.deleteListingById(id);
@@ -237,6 +262,7 @@ public class ListingController {
     }
 
     @Query("getAllListingUniversities")
+    @RolesAllowed({"ListingProvider", "ListingConsumer", "Administrator"})
     @Description("Fetches a list of all universities that listings are assigned to")
     public List<String> getAllListingUniversities(Optional<String> query) {
         LOG.info("getAllListingUniversities");
@@ -248,6 +274,7 @@ public class ListingController {
     }
 
     @Query("getAllListingCompanies")
+    @RolesAllowed({"ListingProvider", "ListingConsumer", "Administrator"})
     @Description("Fetches a list of all companies that listings are assigned to")
     public List<String> getAllListingCompanies(Optional<String> query) {
         LOG.info("getAllListingCompanies");
@@ -255,6 +282,19 @@ public class ListingController {
             return listingService.getAllCompanies(query.get());
         } else {
             return listingService.getAllCompanies();
+        }
+    }
+
+    // Checks if user (identified via JWT) is allowed to edit a listing
+    // Listing Owner is read from DB and compared with JWT Token userID
+    private boolean is_user_allowed_to_access_listing(JsonWebToken jwt, Long listingId) throws GraphQLException {
+        try {
+            Long userId = Long.parseLong(jwt.getClaim("userid").toString());
+            // get owner of listing id in DB
+            Long listing_owner_db = listingService.getListingById(listingId).getOwner().getId();
+            return listing_owner_db.equals(userId);
+        } catch (ServiceException e) {
+            throw new GraphQLException(e.getMessage());
         }
     }
